+++ src/extension.ts
@@ -41,6 +41,8 @@ import { getCodeContextSummary } from "./utils/treeview/getCodeContextSummary";
 import setLoading from "./utils/vscode/setLoading";
 import addToRecommendedCommandHandler from "./utils/commands/addToRecommended";
 import ghHover from "./utils/components/ghHover";
+import slackHover from "./utils/components/slackHover";
+import jiraHover from "./utils/components/jiraHover";
 
 // repo information
 let owner: string | undefined = "";
@@ -304,9 +306,12 @@ export async function activate(context: vscode.ExtensionContext) {
   // update status bar item once at start
   updateStatusBarItem(wmStatusBarItem);
 
-  // create the hover provider
-  let wmHover = hover({ reporter });
-  let githubHover = ghHover({ reporter });
+  // create the hover providers
+
+  hover({ reporter });
+  ghHover({ reporter });
+  slackHover({ reporter });
+  jiraHover({ reporter });
 
   let loginCommandHandler = async () => {
     // Get our PAT session.