+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [2.0.8]
 
 - Added GitHub Hover response
+- Added Slack Hover response
+- Added Jira Hover response
 
 ## [2.0.7]